% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_outline.R
\name{data_outline}
\alias{data_outline}
\title{Compute various descriptive statistics}
\usage{
data_outline(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A data.frame with variables:

\item{N }{The length.}

\item{Mean }{The mean.}

\item{Var }{The variance.}

\item{std_dev }{Standard deviation.}

\item{Median }{The median.}

\item{std_mean }{The standard error of the sample mean.}

\item{CV }{The coefficient of variation.}

\item{CSS }{The corrected sum of squares.}

\item{USS }{The uncorrected sum of squares.}

\item{R }{The extreme difference.}

\item{R1 }{The half extreme difference, or the difference of upper quartile and lower quartile.}

\item{Skewness }{The coefficient of skewness.}

\item{Kurtosis }{The coefficient of kurtosis.}

\item{row.names }{1.}
}
\description{
Compute various descriptive statistics of x, such as mean, median, skewness, and kurtosis, etc.
}
\examples{
x=rnorm(10, mean = 1, sd = 0.2); x
data_outline(x)

}
\references{
Zhang, Y. Y., Wei, Y. (2013), One and two samples using only an R funtion,
\doi{10.2991/asshm-13.2013.29}.
}
\author{
Ying-Ying Zhang (Robert) \email{robertzhangyying@qq.com}
}
