% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneTwoSamples-package.R
\docType{package}
\name{OneTwoSamples-package}
\alias{OneTwoSamples}
\alias{OneTwoSamples-package}
\title{OneTwoSamples: Deal with One and Two (Normal) Samples}
\description{
In this package, we introduce an R function \code{one_two_sample()} which
can deal with one and two (normal) samples, Ying-Ying Zhang, Yi Wei (2012),
\doi{10.2991/asshm-13.2013.29}. For one normal sample x, the function
reports descriptive statistics, plot, interval estimation and test of
hypothesis of x. For two normal samples x and y, the function reports
descriptive statistics, plot, interval estimation and test of hypothesis of
x and y, respectively. It also reports interval estimation and test of
hypothesis of mu1-mu2 (the difference of the means of x and y) and
sigma1^2/sigma2^2 (the ratio of the variances of x and y), tests whether x
and y are from the same population, finds the correlation coefficient of x
and y if x and y have the same length.
}
\details{
The most important functions are \code{\link{one_two_sample}} and
\code{\link{one_sample}}.
}
\references{
Zhang, Y. Y., Wei, Y. (2013), One and two samples using only an
R funtion, \doi{10.2991/asshm-13.2013.29}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fbertran/OneTwoSamples}
  \item \url{https://fbertran.github.io/OneTwoSamples/}
  \item Report bugs at \url{https://github.com/fbertran/OneTwoSamples/issues}
}

}
\author{
\strong{Maintainer}: Frederic Bertrand \email{frederic.bertrand@lecnam.net} (\href{https://orcid.org/0000-0002-0837-8281}{ORCID})

Authors:
\itemize{
  \item Ying-Ying Zhang (Robert)
}

}
\keyword{internal}
